+++ discordia/classes/discord/server.lua
@@ -16,7 +16,6 @@ function Server:__init(data, client)
 
 	self.large = data.large -- boolean
 	self.joinedAt = data.joinedAt -- string
-	self.memberCount = data.memberCount -- number
 
 	if self.large then client.websocket:requestGuildMembers(self.id) end